Change condition to find config function in base Convention
The condition was for the passed `operation` to be an instance of `Operation`. However, `Operation` is an enum defined in microcosm-flask; since Enums cannot be extended, it makes it impossible for a user of `microcosm-flask` to define their own `Operation`-like enum and use its members in a endpoint mapping.