Separate route path construction into its own component
Path construction is difficult to modify today, in part because the `make_path` function
is not exposed and is not customizable.

This PR takes the first step of putting path construction into a graph component, adding
more exhaustive tests, and improving (read: regressing) a few behaviors. Customization
is possible via a `prefix` override parameter, but the builder cannot be swapped out yet.

There are two backwards imcompatible changes here:

 - `Namespace.prefix` is no longer used for the namespace's own path calculation; it is instead
   used by the `RoutePathBuilder` component. By inspection, there are no obvious usages of this
   parameter in our code any more (we use `version` instead), so this should be safe.

 - Trailing slashes are removed from paths. This means that the current discovery convention
   that handles `/api/` will handle `/api` instead. No code should depend on this endpoint and
   the new usage should be better for humans.