Fix swagger def generation for no version/no prefix case
**Why?**
Though it's not something we've done before, one could want to have non-namespaced paths, i.e. all paths under the root without a prefix or version number.
One example is when paths are imposed by a third-party tool that will contact our API on a fixed path (e.g. building a container for AWS SageMaker).

**What?**
The way we generate paths in the OpenAPI definition assumes that the combination of API version + prefix (the "base path") was more than `"/"`, leaving each path in the definition without a leading underscore (which makes the OpenAPI definition invalid).
- Fix paths in the OpenAPI definition to have a leading slash when the API prefix and version are empty strings.
- Add test for that particular case.