Fix swagger def generation for no version/no prefix case (#204)
**Why?**
Though it's not something we've done before, one could want to have non-namespaced paths, i.e. all paths under the root without a prefix or version number.
One example is when paths are imposed by a third-party tool that will contact our API on a fixed path (e.g. building a container for AWS SageMaker).

Since a path suffix has a leading /, when the API prefix and version number are both empty we'd expect the base path to equal "". However, OpenAPI requires the base path to minimally be "/", which requires a special case to generate a proper OpenAPI definition.

**What?**
- Fix paths in the OpenAPI definition to have a leading slash when the API prefix and version are empty strings.
- Add test for that particular case.