Allow audit logging at debug. (#215)
* Allow audit logging at debug.

When we set out to add audit logging to microcosm we had a more naive
definition of audit logs. I think what we call "audit logs" are just
"request logs" and should be revisted later. For now, we need the
ability to turn some of these automatic logs off by lowering the logging
level to clean up our logs.