Add memory profiling support.
Why? We have excellent CPU profiling support, but no memory support.
Recently we enocuntered a memory leak integrating another third party
library and found ourselves blind to debugging the leak in our
environment.

This tooling is primitive but should be sufficient for identifying leaks
in the application. This approach is copied from:

https://superuser.blog/detect-memory-leak-python/

which uses a built-in python memory tracking library: `tracemalloc`.