Make span names better for tracing.
Why? We were using this and found that the span names that we were
getting we including a lot of information we didn't want, e.g. query
params, etc and in some cases could be very long. This had consequences
for memory usage due to how `jaeger` tries to save information for
inter-process communication.