Introduce associated schemas
Currently, schema definitions that we expose in the swagger convention are all tied to a route.
Here we introduce the concept of an "associated schema", which is derived from another resource but not tied to a route.

One use case we have is for pubsub: our pubsub messages are closely tied to our route resources; however we don't necessarily send all of a resource fields in a pubsub message. The mechanism introduced here will allow us to use swagger schemas to define those message schemas.

- Introduce `add_associated_schema` decorator.
- Use a schema's associated schemas when generating swagger definitions.