Fix building OpenAPI definition for integer-valued ids
**Why?**
The way we build OpenAPI definitions for flask routes breaks on non-string valued ids.
In particular, it breaks for integer-valued ids. This is because we use werkzeug functions to build paths in the OpenAPI definitions; using parameter names instead of actual values. This works fine for string-valued parameters, but not in other cases.
In particular, it breaks for integer-valued IDs.

**What?**
Use placeholder integers to build URIs using werkzeug functions, and replace placeholders with param names afterwards.