Support data_key usage for field names for generated swagger
The recommended method of specifying serialized names in marshmallow is
to use the `data_key` argument when specifying a field, instead of `attribute`
(which otherwise leads to the awkward inclusion of camelCased field names,
which is less than ideal).

Unfortunately, this otherwise breaks existing swagger definition
generation, as it didn't take this field into account.

This will use an existing `data_key` field if it exists as a name, but
otherwise fallback to the predefined name.

https://marshmallow.readthedocs.io/en/stable/quickstart.html#specifying-serialization-deserialization-keys