GLOB-44246 Improve swagger endpoint performance
As the number of endpoints grows, the time taken for this endpoint grows
non-linearly. This is due to the fact that during `add_paths`, `url_for`
performs what is essentially a full scan across every possible endpoint
that the app is already aware of, performing a relatively expensive
diff on each turn.

While this only occurs on BuildError, it is frequently raised when the endpoint
in question requires any path argument, which ends up being many of them
(retrieve/delete/update/etc.).

We can sidestep that error by performing a preflight operation (which
`url_for` is already end up doing under the hood) that pre-fetches the
expected arguments used in any given path, and passes that in when
constructing the path.

Also add an extra flag to enable schema validation, as this can
frequently eat up a non-trivial amount of time. This will, however, push
the burden of validating schemas onto the project itself (e.g. via an
additional unit test)