Enable optional health checks
Adds the ability to cordon off essential health checks from other
"diagnostic" type of checks.

The health endpoint is intended to be quick, non-intrusive, and accurate
with regards to providing a minimal guarantee of functionality for a
service. However, because there was only one avenue for adding any type of
"status" check here, the endpoint has trended to be slower over time.

This will allow those slower checks to be seen on-demand, but not
interfere with the default checks used by monitoring systems.