Update landing page to use optional full health check param (#313)
- update landing page to show optional full health check data

Why?

The default health endpoint returns a minimum set of health data. This
is generally used to make sure a service is "up".

The landing page is meant to be a human consumable resource showing
all health and config data. Since this view is meant for humans to see
a complete view of a service it makes sense to show full health check
data.